test(daemon): pin the teardown lazy seam with an eager-closure guard
Extracts the AST-based eager-import-closure walker from
cli-startup-import-closure.test.ts into eager-import-closure.fixtures.ts and
adds session-teardown-import-closure.test.ts, which fails if either android
cleanup helper returns to a static import. Observed red against main's
static imports before green.

Reverts the apple perf-xctrace lazy hunk: the Apple unit test mocks the
cleanup module, so it could not prove the dynamic chunk loads on the shipped
path.
